move TestFunctionalIndex to functional tests package
